+++ repo/src/main/java/com/parashift/onlyoffice/Download.java
@@ -37,10 +37,14 @@ public class Download extends AbstractWebScript {
     @Autowired
     MimetypeService mimetypeService;
 
+    @Autowired
+    Util util;
+
     @Override
     public void execute(WebScriptRequest request, WebScriptResponse response) throws IOException {
         if (request.getParameter("nodeRef") != null) {
-            if (jwtManager.jwtEnabled()) {
+            String zipParam = request.getParameter("wjc");
+            if (jwtManager.jwtEnabled() && zipParam == null) {
                 String jwth = jwtManager.getJwtHeader();
                 String header = request.getHeader(jwth);
                 String token = (header != null && header.startsWith("Bearer ")) ? header.substring(7) : header;
@@ -59,13 +63,27 @@ public class Download extends AbstractWebScript {
             if (permissionService.hasPermission(nodeRef, PermissionService.READ) != AccessStatus.ALLOWED) {
                 throw new AccessDeniedException("Access denied. You do not have the appropriate permissions to perform this operation");
             }
-
+            if (zipParam != null) {
+                NodeRef parentVersionNode = nodeService.getParentAssocs(nodeRef).get(0).getParentRef();
+                if (permissionService.hasPermission(parentVersionNode, PermissionService.READ) != AccessStatus.ALLOWED) {
+                    throw new AccessDeniedException("Access denied. You do not have the appropriate permissions to perform this operation");
+                }
+            }
             ContentReader reader = contentService.getReader(nodeRef, ContentModel.PROP_CONTENT);
             Map<QName, Serializable> properties = nodeService.getProperties(nodeRef);
 
             String name = (String) properties.get(ContentModel.PROP_NAME);
             String docExt = name.substring(name.lastIndexOf(".") + 1).trim().toLowerCase();
 
+            String editorUrl = util.getEditorUrl();
+            if (editorUrl.endsWith("/")) {
+                editorUrl = editorUrl.substring(0, editorUrl.length() - 1);
+            }
+
+            response.setHeader("Access-Control-Allow-Origin", editorUrl);
+            response.setHeader("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, Authorization");
+            response.setHeader("Access-Control-Allow-Credentials", "true");
+            response.setHeader("Access-Control-Allow-Methods", "GET,POST,PUT,DELETE,OPTIONS,HEAD");
             response.setHeader("Content-Length", String.valueOf(reader.getSize()));
             response.setHeader("Content-Type", mimetypeService.getMimetype(docExt));
             response.setHeader("Content-Disposition", "attachment; filename*=UTF-8\'\'" + URLEncoder.encode(name));