+++ repo/src/main/java/com/parashift/onlyoffice/Prepare.java
@@ -1,17 +1,15 @@
 package com.parashift.onlyoffice;
 
 import org.alfresco.model.ContentModel;
-import org.alfresco.service.cmr.repository.ContentService;
-import org.alfresco.service.cmr.repository.ContentWriter;
-import org.alfresco.service.cmr.repository.MimetypeService;
-import org.alfresco.service.cmr.repository.NodeRef;
-import org.alfresco.service.cmr.repository.NodeService;
-import org.alfresco.service.cmr.security.AccessStatus;
-import org.alfresco.service.cmr.security.PermissionService;
-import org.alfresco.service.namespace.NamespaceService;
-import org.alfresco.service.namespace.QName;
 import org.alfresco.repo.i18n.MessageService;
 import org.alfresco.repo.security.authentication.AuthenticationUtil;
+import org.alfresco.service.cmr.repository.*;
+import org.alfresco.service.cmr.security.AccessStatus;
+import org.alfresco.service.cmr.security.AuthenticationService;
+import org.alfresco.service.cmr.security.PermissionService;
+import org.alfresco.service.cmr.version.VersionService;
+import org.alfresco.service.namespace.NamespaceService;
+import org.alfresco.service.namespace.QName;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -66,6 +64,12 @@ public class Prepare extends AbstractWebScript {
     @Autowired
     UtilDocConfig utilDocConfig;
 
+    @Autowired
+    AuthenticationService authenticationService;
+
+    @Autowired
+    VersionService versionService;
+
     @Override
     public void execute(WebScriptRequest request, WebScriptResponse response) throws IOException {
         mesService.registerResourceBundle("alfresco/messages/prepare");
@@ -102,8 +106,15 @@ public class Prepare extends AbstractWebScript {
                         if (pathLocale == null) pathLocale = Util.PathLocale.get("en");
                     }
 
-                    InputStream in = getClass().getResourceAsStream("/newdocs/" + pathLocale + "/new." + ext);
-
+                    InputStream in;
+                    NodeRef parentNodeRef = null;
+                    if(request.getParameter("parentNodeRef") != null){
+                        parentNodeRef = new NodeRef(request.getParameter("parentNodeRef"));
+                        in = contentService.getReader(parentNodeRef, ContentModel.PROP_CONTENT).getContentInputStream();
+                    }
+                    else {
+                        in = getClass().getResourceAsStream("/newdocs/" + pathLocale + "/new." + ext);
+                    }
                     writer.putContent(in);
 
                     util.ensureVersioningEnabled(nodeRef);
@@ -158,6 +169,10 @@ public class Prepare extends AbstractWebScript {
                 responseJson.put("mime", mimetypeService.getMimetype(docExt));
                 responseJson.put("demo", configManager.demoActive());
                 responseJson.put("documentType", documentType);
+                responseJson.put("historyUrl", util.getHistoryUrl(nodeRef));
+                responseJson.put("share", util.getShareUrl());
+                responseJson.put("favorite", util.getFavouriteUrl(nodeRef));
+                responseJson.put("ticket",  authenticationService.getCurrentTicket());
 
                 logger.debug("Sending JSON prepare object");
                 logger.debug(responseJson.toString(3));